Added controllers and middlewares, but not all of them. Initial data is getting filled out and tests will be written after a better division between middlewares and controllers can be decided.

+++ models/flourishflora.js
@@ -2,7 +2,7 @@ const flourishflora = (connection, Sequelize, Samples) => {
   return connection.define('flourishflora', {
     id: { type: Sequelize.INTEGER, autoIncrement: true, primaryKey: true },
     sampleId: { type: Sequelize.INTEGER, references: { model: Samples, key: 'id' } },
-    producerCoefficient: { type: Sequelize.INTEGER }
+    producerCoefficient: { type: Sequelize.DECIMAL },
   }, {
     defaultScope: {
       attributes: { exclude: ['deletedAt'] }
